requirements.txt
# OpenMTSciEd 核心依赖包

# --- 后端框架 ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
PyJWT>=2.8.0
python-multipart>=0.0.6

# --- 数据处理与解析 ---
pandas>=2.0.0
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# --- 数据库 ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
alembic>=1.12.0
python-dotenv>=1.0.0

# --- 知识图谱 ---
neo4j>=5.0.0

# --- AI 辅助 (轻量级) ---
langchain>=0.0.300
transformers>=4.35.0

# --- 测试 ---
pytest>=7.4.0
pytest-asyncio>=0.21.0
